Minor Pico-8 external launch changes

+++ script/launch/ext-pico8.sh
@@ -21,12 +21,11 @@ SETUP_SDL_ENVIRONMENT
 P8_BIN="pico8_64"
 SET_VAR "system" "foreground_process" "$P8_BIN"
 
-GPTOKEYB="$(GET_VAR "device" "storage/rom/mount")/MUOS/emulator/gptokeyb/gptokeyb2"
-EMUDIR="$(GET_VAR "device" "storage/rom/mount")/MUOS/emulator/pico8"
-
 # Set appropriate SDL controller file
 SDL_GAMECONTROLLERCONFIG_FILE="/opt/muos/device/control/gamecontrollerdb_modern.txt"
 export SDL_GAMECONTROLLERCONFIG_FILE
+
+EMUDIR="$(GET_VAR "device" "storage/rom/mount")/MUOS/emulator/pico8"
 
 # First look for emulator in BIOS directory, which allows it to follow the
 # user's storage preference. Fall back on the old path for compatibility.
@@ -34,9 +33,7 @@ export SDL_GAMECONTROLLERCONFIG_FILE
 EMU="/run/muos/storage/bios/pico8/$P8_BIN"
 if [ ! -f "$EMU" ]; then
 	EMU="/run/muos/storage/bios/pico-8/$P8_BIN"
-	if [ ! -f "$EMU" ]; then
-		EMU="$EMUDIR/$P8_BIN"
-	fi
+	[ ! -f "$EMU" ] && EMU="$EMUDIR/$P8_BIN"
 fi
 
 # Did the user select standard or Pixel Perfect scaler?
@@ -55,61 +52,58 @@ cd "$EMUDIR" || exit
 
 F_DIR="$(dirname "$FILE")"
 
+GPTOKEYB="$(GET_VAR "device" "storage/rom/mount")/MUOS/emulator/gptokeyb/gptokeyb2"
+$GPTOKEYB "./$P8_BIN" -c "./pico8.gptk" &
+
 if [ "$NAME" = "Splore" ]; then
-	SDL_ASSERT=always_ignore \
-		$GPTOKEYB "./$P8_BIN" -c "./pico8.gptk" &
-	PATH="$EMUDIR:$PATH" \
-		HOME="$EMUDIR" \
-		"$EMU" $PICO_FLAGS -root_path "$F_DIR" -splore
+	SDL_ASSERT=always_ignore PATH="$EMUDIR:$PATH" HOME="$EMUDIR" "$EMU" $PICO_FLAGS -root_path "$F_DIR" -splore
 else
-	SDL_ASSERT=always_ignore \
-		$GPTOKEYB "./$P8_BIN" -c "./pico8.gptk" &
-	PATH="$EMUDIR:$PATH" \
-		HOME="$EMUDIR" \
-		"$EMU" $PICO_FLAGS -root_path "$F_DIR" -run "$FILE"
+	SDL_ASSERT=always_ignore PATH="$EMUDIR:$PATH" HOME="$EMUDIR" "$EMU" $PICO_FLAGS -root_path "$F_DIR" -run "$FILE"
 fi
 
-kill -9 "$(pidof $P8_BIN)" "$(pidof gptokeyb2)"
+killall -9 "$(pidof $P8_BIN)" "$(pidof gptokeyb2)"
 
 /opt/muos/script/mux/track.sh "$NAME" "$CORE" "$FILE" stop
 
 unset SDL_HQ_SCALER SDL_ROTATION SDL_BLITTER_DISABLED
 
-# SAVE THE FAVOURITES CHARLIE!
-# Grab the PICO-8 ROM Folder
-STORAGE_DIR=${FILE%/*}
-
 FAVOURITE="/run/muos/storage/save/pico8/favourites.txt"
-CART_DIR="/run/muos/storage/save/pico8/bbs"
-BOXART_DIR="/run/muos/storage/info/catalogue/PICO-8/box"
+if [ -e $FAVOURITE ]; then
+	# SAVE THE FAVOURITES CHARLIE!
+	# Grab the PICO-8 ROM Folder
+	STORAGE_DIR=${FILE%/*}
 
-# TODO: Work out what these other fields mean?! (maybe useful?)
-while IFS='|' read -r _ RAW_NAME _ _ _ _ GOOD_NAME; do
-	[ -z "$GOOD_NAME" ] || [ -z "$RAW_NAME" ] && continue
-	RAW_NAME=$(echo "$RAW_NAME" | sed 's/^[[:space:]]*//;s/[[:space:]]*$//;s/[[:space:]]\+//g')
-	GOOD_NAME=$(echo "$GOOD_NAME" | sed -E 's/.*\|//;s/^[[:space:]]+|[[:space:]]+$//;s/\b(.)/\u\1/g' | tr -d ':')
+	CART_DIR="/run/muos/storage/save/pico8/bbs"
+	BOXART_DIR="/run/muos/storage/info/catalogue/PICO-8/box"
 
-	P8_SRC_EXT="p8.png"
-	DEST_EXT="p8"
-	PNG_EXT="png"
+	# TODO: Work out what these other fields mean?! (maybe useful?)
+	while IFS='|' read -r _ RAW_NAME _ _ _ _ GOOD_NAME; do
+		[ -z "$GOOD_NAME" ] || [ -z "$RAW_NAME" ] && continue
+		RAW_NAME=$(echo "$RAW_NAME" | sed 's/^[[:space:]]*//;s/[[:space:]]*$//;s/[[:space:]]\+//g')
+		GOOD_NAME=$(echo "$GOOD_NAME" | sed -E 's/.*\|//;s/^[[:space:]]+|[[:space:]]+$//;s/\b(.)/\u\1/g' | tr -d ':')
 
-	FAV_FILE=""
+		P8_SRC_EXT="p8.png"
+		DEST_EXT="p8"
+		PNG_EXT="png"
 
-	for DIR in "$CART_DIR" "$CART_DIR"/*; do
-		DIR="${DIR%/}"
-		if [ "$(basename "$DIR")" = "labels" ]; then
-			continue
+		FAV_FILE=""
+
+		for DIR in "$CART_DIR" "$CART_DIR"/*; do
+			DIR="${DIR%/}"
+			if [ "$(basename "$DIR")" = "labels" ]; then
+				continue
+			fi
+			if [ -f "$DIR/$RAW_NAME.$P8_SRC_EXT" ]; then
+				FAV_FILE="$DIR/$RAW_NAME.$P8_SRC_EXT"
+				break
+			fi
+		done
+
+		if [ -n "$FAV_FILE" ]; then
+			DEST_FILE="$STORAGE_DIR/$GOOD_NAME.$DEST_EXT"
+			BOXART_FILE="$BOXART_DIR/$GOOD_NAME.$PNG_EXT"
+			cp "$FAV_FILE" "$DEST_FILE"
+			cp "$FAV_FILE" "$BOXART_FILE"
 		fi
-		if [ -f "$DIR/$RAW_NAME.$P8_SRC_EXT" ]; then
-			FAV_FILE="$DIR/$RAW_NAME.$P8_SRC_EXT"
-			break
-		fi
-	done
-
-	if [ -n "$FAV_FILE" ]; then
-		DEST_FILE="$STORAGE_DIR/$GOOD_NAME.$DEST_EXT"
-		BOXART_FILE="$BOXART_DIR/$GOOD_NAME.$PNG_EXT"
-		cp "$FAV_FILE" "$DEST_FILE"
-		cp "$FAV_FILE" "$BOXART_FILE"
-	fi
-done <"$FAVOURITE"
+	done <"$FAVOURITE"
+fi
